+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes for Sprig
 
-## 1.0.0-beta.10 - Unreleased
+## 1.0.0-beta.10 - 2020-08-04
 ### Added
 - Added the ability to override the component trigger using the `trigger` attribute.